+++ DooRiBon/DooRiBon/Resources/Extensions/Date+Extension.swift
@@ -9,23 +9,4 @@ import UIKit
 
 extension Formatter {
     static let date = DateFormatter()
-    
-     // - 함수 실행 getDatesBetweenTwo(from: "2021-07-14",to: "2021-07-20")
-     // - 배열 반환 ["2021-07-14", ""2021-07-15", ""2021-07-16", ..., "2021-07-20"]
-    static func getDatesBetweenTwo(from startDate: String, to date: String) -> [String] {
-        Formatter.date.dateFormat = "yyyy-MM-dd"
-        
-        // 시작 날짜와 끝 날짜를 date 형식으로 변환
-        guard var startDate = Formatter.date.date(from: startDate) else { return [] }
-        guard let endDate = Formatter.date.date(from: date) else { return [] }
-
-        var dates: [String] = []
-        
-        // while date less than or equal to end date
-        while startDate <= endDate {
-            dates.append( Formatter.date.string(from: startDate))
-            startDate = Calendar.current.date(byAdding: .day, value: 1, to: startDate)!
-        }
-        return dates
-    }
 }
